Create a separate plugin file (as in the mutations PR)
